+++ .travis.yml
@@ -10,6 +10,6 @@ install:
   - pip install -r test-requirements.txt -r requirements.txt
 
 script:
-  - flake8 --config ./lint-configs/python/.flake8 st2auth_enterprise_ldap_backend/
-  - pylint -E --rcfile=./lint-configs/python/.pylintrc st2auth_enterprise_ldap_backend/
-  - nosetests -sv --with-coverage --cover-package=st2auth_enterprise_ldap_backend tests/unit
+  - flake8 --config ./lint-configs/python/.flake8 st2auth_ldap/
+  - pylint -E --rcfile=./lint-configs/python/.pylintrc st2auth_ldap/
+  - nosetests -sv --with-coverage --cover-package=st2auth_ldap tests/unit
