+++ results.txt
@@ -1,0 +1,3 @@
+# Automatically generated
+# duptype id depth size device inode priority name
+# end of file
