+++ tests/integration/test_openldap.py
@@ -15,7 +15,7 @@
 import os
 import unittest2
 
-from st2auth_enterprise_ldap_backend import ldap_backend
+from st2auth_ldap import ldap_backend
 
 
 class OpenLDAPAuthenticationTest(unittest2.TestCase):
