+++ tests/unit/test_config.py
@@ -18,7 +18,7 @@ import mock
 import unittest2
 import uuid
 
-from st2auth_enterprise_ldap_backend import ldap_backend
+from st2auth_ldap import ldap_backend
 
 
 LDAP_HOST = '127.0.0.1'
