+++ tox.ini
@@ -7,5 +7,5 @@ deps = -r{toxinidir}/requirements.txt
 commands = nosetests -sv tests/unit
 
 [testenv:lint]
-commands = flake8 --config ./lint-configs/python/.flake8 st2auth_enterprise_ldap_backend/
-           pylint -E --rcfile=./lint-configs/python/.pylintrc st2auth_enterprise_ldap_backend/
+commands = flake8 --config ./lint-configs/python/.flake8 st2auth_ldap/
+           pylint -E --rcfile=./lint-configs/python/.pylintrc st2auth_ldap/
